+++ docs/changelog.md
@@ -1,5 +1,7 @@
 2/7/2023: Karl
 - abstracted command line argument parse and calls to a new function, CompFiles_FileOpenFromCLIArgs, which also generates the Temp file.
+- created Tokens_GetName, the tokens Map, the tokens enum.
+- created a state transition table as a 3d array for a keyword recognizer Token_RecognizeKeyword. Used excel to design the table; .xlsx is in the /docs folder.